Add _Plotter and _XPostitions classes
