+++ statannotations/Annotator.py
@@ -4,14 +4,12 @@ from typing import List
 import matplotlib.pyplot as plt
 import matplotlib.transforms as mtransforms
 import numpy as np
-import seaborn as sns
 from matplotlib import lines
-from matplotlib.collections import PathCollection
 from matplotlib.font_manager import FontProperties
-from matplotlib.patches import Rectangle
 
 from statannotations.PValueFormat import PValueFormat, \
     CONFIGURABLE_PARAMETERS as PVALUE_CONFIGURABLE_PARAMETERS
+from statannotations._Plotter import _Plotter
 from statannotations.stats.ComparisonsCorrection import \
     get_validated_comparisons_correction
 from statannotations.stats.StatResult import StatResult
@@ -19,8 +17,7 @@ from statannotations.stats.StatTest import StatTest
 from statannotations.stats.test import apply_test, IMPLEMENTED_TESTS
 from statannotations.stats.utils import check_alpha, check_pvalues, \
     check_num_comparisons, get_num_comparisons
-from statannotations.utils import check_is_in, remove_null, check_not_none, \
-    check_box_pairs_in_data, check_order_in_data, render_collection
+from statannotations.utils import check_is_in, render_collection
 
 CONFIGURABLE_PARAMETERS = [
     'alpha',
@@ -40,8 +37,6 @@ CONFIGURABLE_PARAMETERS = [
     'verbose',
 ]
 
-IMPLEMENTED_PLOTTERS = ['barplot', 'boxplot', 'stripplot', 'swarmplot',
-                        'violinplot']
 
 _DEFAULT_VALUES = {
     "alpha": 0.05,
@@ -60,12 +55,6 @@ _DEFAULT_VALUES = {
     "line_width": 1.5,
     "custom_annotations": None
 }
-
-
-def check_implemented_plotters(plot):
-    if plot not in IMPLEMENTED_PLOTTERS:
-        raise NotImplementedError(
-            f"Only {render_collection(IMPLEMENTED_PLOTTERS)} are supported.")
 
 
 class Annotator:
@@ -105,13 +94,8 @@ class Annotator:
         self.box_pairs = box_pairs
         self.ax = ax
 
-        basic_plot = self.get_basic_plot(ax, box_pairs, plot, data, x, y, hue,
-                                         order, hue_order, **plot_params)
-
-        self.y_stack_arr, self.box_struct_pairs, self.fig = basic_plot
-
-        self._reordering = None
-        self._sort_box_struct_pairs()
+        self._plotter = _Plotter(ax, box_pairs, plot, data, x, y, hue,
+                                 order, hue_order, **plot_params)
 
         self._test = None
         self.perform_stat_test = None
@@ -140,35 +124,6 @@ class Annotator:
         self.OFFSET_FUNCS = {"inside": Annotator._get_offsets_inside,
                              "outside": Annotator._get_offsets_outside}
 
-    @staticmethod
-    def get_basic_plot(
-            ax, box_pairs, plot='boxplot', data=None, x=None, y=None, hue=None,
-            order=None, hue_order=None, **plot_params):
-        check_implemented_plotters(plot)
-        check_not_none("box_pairs", box_pairs)
-        check_order_in_data(data, x, order)
-        check_box_pairs_in_data(box_pairs, data, x, hue, hue_order)
-
-        plotter = Annotator._get_plotter(plot, x, y, hue, data, order,
-                                         hue_order, **plot_params)
-
-        box_names, labels = Annotator._get_box_names_and_labels(plotter)
-
-        fig = plt.gcf()
-        ymaxes = Annotator._generate_ymaxes(ax, fig, plotter, box_names)
-        box_structs = Annotator._get_box_structs(
-            plotter, box_names, labels, ymaxes)
-
-        box_struct_pairs = Annotator._get_box_struct_pairs(
-            box_pairs, box_structs)
-
-        y_stack_arr = np.array(
-            [[box_struct['x'], box_struct['ymax'], 0]
-             for box_struct in box_structs]
-        ).T
-
-        return y_stack_arr, box_struct_pairs, fig
-
     def new_plot(self, ax, box_pairs=None, plot='boxplot', data=None, x=None,
                  y=None, hue=None, order=None, hue_order=None, **plot_params):
         self.ax = ax
@@ -176,18 +131,26 @@ class Annotator:
         if box_pairs is None:
             box_pairs = self.box_pairs
 
-        basic_plot = self.get_basic_plot(ax, box_pairs, plot, data, x, y, hue,
-                                         order, hue_order, **plot_params)
-
-        self.y_stack_arr, self.box_struct_pairs, self.fig = basic_plot
-
-        self._sort_box_struct_pairs()
+        self._plotter = _Plotter(ax, box_pairs, plot, data, x, y, hue,
+                                 order, hue_order, **plot_params)
 
         self.line_offset = None
         self.line_offset_to_box = None
         self.perform_stat_test = None
 
         return self
+
+    @property
+    def _y_stack_arr(self):
+        return self._plotter.y_stack_arr
+
+    @property
+    def fig(self):
+        return self._plotter.fig
+
+    @property
+    def box_struct_pairs(self):
+        return self._plotter.box_struct_pairs
 
     def reset_configuration(self):
 
@@ -215,7 +178,7 @@ class Annotator:
         if self.verbose:
             self.print_pvalue_legend()
 
-        ax_to_data = Annotator._get_transform_func(self.ax, 'ax_to_data')
+        ax_to_data = self._plotter.get_transform_func('ax_to_data')
 
         self.validate_test_short_name()
         annotations = self.custom_annotations or self.annotations
@@ -228,8 +191,8 @@ class Annotator:
                                 orig_ylim=orig_ylim)
 
         # reset transformation
-        y_stack_max = max(self.y_stack_arr[1, :])
-        ax_to_data = Annotator._get_transform_func(self.ax, 'ax_to_data')
+        y_stack_max = max(self._y_stack_arr[1, :])
+        ax_to_data = self._plotter.get_transform_func('ax_to_data')
         ylims = ([(0, 0), (0, max(1.04 * y_stack_max, 1))]
                  if self.loc == 'inside'
                  else [(0, 0), (0, 1)])
@@ -366,6 +329,10 @@ class Annotator:
         return None
 
     @property
+    def _reordering(self):
+        return self._plotter.reordering
+
+    @property
     def alpha(self):
         return self._alpha
 
@@ -455,8 +422,7 @@ class Annotator:
             test_result_list.append(result)
 
         # Perform other types of correction methods for multiple testing
-        Annotator._apply_comparisons_correction(self.comparisons_correction,
-                                                test_result_list)
+        self._apply_comparisons_correction(test_result_list)
 
         return test_result_list
 
@@ -483,8 +449,8 @@ class Annotator:
 
         # Find y maximum for all the y_stacks *in between* box1 and box2
         i_ymax_in_range_x1_x2 = xi1 + np.nanargmax(
-            self.y_stack_arr[1, np.where((x1 <= self.y_stack_arr[0, :])
-                                         & (self.y_stack_arr[0, :] <= x2))])
+            self._y_stack_arr[1, np.where((x1 <= self._y_stack_arr[0, :])
+                                          & (self._y_stack_arr[0, :] <= x2))])
 
         y = self._get_y_for_pair(i_ymax_in_range_x1_x2)
 
@@ -499,13 +465,13 @@ class Annotator:
         line_x, line_y = zip(*points)
         self._plot_line(line_x, line_y)
 
+        ann = self.ax.annotate(
+            text, xy=(np.mean([x1, x2]), line_y[2]),
+            xytext=(0, self.text_offset), textcoords='offset points',
+            xycoords='data', ha='center', va='bottom',
+            fontsize=self._pvalue_format.fontsize, clip_on=False,
+            annotation_clip=False)
         if text is not None:
-            ann = self.ax.annotate(
-                text, xy=(np.mean([x1, x2]), line_y[2]),
-                xytext=(0, self.text_offset), textcoords='offset points',
-                xycoords='data', ha='center', va='bottom',
-                fontsize=self._pvalue_format.fontsize, clip_on=False,
-                annotation_clip=False)
             ann_list.append(ann)
             plt.draw()
             self.ax.set_ylim(orig_ylim)
@@ -516,205 +482,16 @@ class Annotator:
 
         # Fill the highest y position of the annotation into the y_stack array
         # for all positions in the range x1 to x2
-        self.y_stack_arr[1,
-                         (x1 <= self.y_stack_arr[0, :])
-                         & (self.y_stack_arr[0, :] <= x2)] = y_top_annot
+        self._y_stack_arr[1,
+                          (x1 <= self._y_stack_arr[0, :])
+                          & (self._y_stack_arr[0, :] <= x2)] = y_top_annot
 
         # Increment the counter of annotations in the y_stack array
-        self.y_stack_arr[2, xi1:xi2 + 1] += 1
+        self._y_stack_arr[2, xi1:xi2 + 1] += 1
 
     def _update_y_for_loc(self):
         if self._loc == 'outside':
-            self.y_stack_arr[1, :] = 1
-
-    @staticmethod
-    def _get_box_x_position(b_plotter, box_name):
-        """
-        box_name can be either a name "cat" or a tuple ("cat", "hue")
-        """
-        if b_plotter.plot_hues is None:
-            cat = box_name
-            hue_offset = 0
-        else:
-            cat = box_name[0]
-            hue_level = box_name[1]
-            hue_offset = b_plotter.hue_offsets[
-                b_plotter.hue_names.index(hue_level)]
-
-        group_pos = b_plotter.group_names.index(cat)
-        box_pos = group_pos + hue_offset
-        return box_pos
-
-    @staticmethod
-    def _get_xpos_location(pos, xranges):
-        """
-        Finds the x-axis location of a categorical variable
-        """
-        for xrange in xranges:
-            if (pos >= xrange[0]) & (pos <= xrange[1]):
-                return xrange[2]
-
-    @staticmethod
-    def _get_ypos_for_line2d_or_rectangle(
-            xpositions, data_to_ax, ax, fig, child):
-        xunits = (max(list(xpositions.keys())) + 1) / len(xpositions)
-        xranges = {(pos - xunits / 2, pos + xunits / 2, pos): box_name
-                   for pos, box_name in xpositions.items()}
-        box = ax.transData.inverted().transform(
-            child.get_window_extent(fig.canvas.get_renderer()))
-
-        if (box[:, 0].max() - box[:, 0].min()) > 1.1 * xunits:
-            return None, None
-        raw_xpos = np.round(box[:, 0].mean(), 1)
-        xpos = Annotator._get_xpos_location(raw_xpos, xranges)
-        if xpos not in xpositions:
-            return None, None
-        xname = xpositions[xpos]
-        ypos = box[:, 1].max()
-        ypos = data_to_ax.transform((0, ypos))[1]
-
-        return xname, ypos
-
-    @staticmethod
-    def _get_ypos_for_path_collection(xpositions, data_to_ax, child):
-        ymax = child.properties()['offsets'][:, 1].max()
-        xpos = float(np.round(np.nanmean(
-            child.properties()['offsets'][:, 0]), 1))
-
-        xname = xpositions[xpos]
-        ypos = data_to_ax.transform((0, ymax))[1]
-
-        return xname, ypos
-
-    @staticmethod
-    def _generate_ymaxes(ax, fig, plotter, box_names):
-        """
-        given box plotter and the names of two categorical variables,
-        returns highest y point drawn between those two variables before
-        annotations.
-
-        The highest data point is often not the highest item drawn
-        (eg, error bars and/or bar charts).
-        """
-        xpositions = {
-            np.round(Annotator._get_box_x_position(plotter, box_name),
-                     1): box_name
-            for box_name in box_names}
-
-        ymaxes = {name: 0 for name in box_names}
-
-        data_to_ax = Annotator._get_transform_func(ax, 'data_to_ax')
-        # noinspection PyProtectedMember
-        if isinstance(plotter, sns.categorical._ViolinPlotter):
-            ymaxes = Annotator._get_ymaxes_violin(plotter, ymaxes, data_to_ax)
-
-        else:
-            for child in ax.get_children():
-
-                xname, ypos = Annotator._get_child_ypos(
-                    child, ax, fig, xpositions, data_to_ax)
-
-                if ypos is not None and ypos > ymaxes[xname]:
-                    ymaxes[xname] = ypos
-
-        return ymaxes
-
-    @staticmethod
-    def _get_child_ypos(child, ax, fig, xpositions, data_to_ax):
-        if ((type(child) == PathCollection)
-                and len(child.properties()['offsets'])):
-            return Annotator._get_ypos_for_path_collection(
-                xpositions, data_to_ax, child)
-
-        elif (type(child) == lines.Line2D) or (type(child) == Rectangle):
-            return Annotator._get_ypos_for_line2d_or_rectangle(
-                xpositions, data_to_ax, ax, fig, child)
-
-        return None, None
-
-    @staticmethod
-    def _get_ymaxes_violin(plotter, ymaxes, data_to_ax):
-        for group_idx, group_name in enumerate(plotter.group_names):
-            if plotter.hue_names:
-                for hue_idx, hue_name in enumerate(plotter.hue_names):
-                    ypos = max(plotter.support[group_idx][hue_idx])
-                    ymaxes[(group_name, hue_name)] = \
-                        data_to_ax.transform((0, ypos))[1]
-            else:
-                ypos = max(plotter.support[group_idx])
-                ymaxes[group_name] = data_to_ax.transform((0, ypos))[1]
-        return ymaxes
-
-    @staticmethod
-    def _get_group_data(b_plotter, cat):
-        index = b_plotter.group_names.index(cat)
-
-        return b_plotter.plot_data[index], index
-
-    @staticmethod
-    def _get_box_data_with_hue(b_plotter, box_name):
-        cat = box_name[0]
-        group_data, index = Annotator._get_group_data(b_plotter, cat)
-
-        hue_level = box_name[1]
-        hue_mask = b_plotter.plot_hues[index] == hue_level
-
-        return group_data[hue_mask]
-
-    @staticmethod
-    def _get_box_data_without_hue(b_plotter, box_name):
-        cat = box_name
-        group_data, _ = Annotator._get_group_data(b_plotter, cat)
-
-        return group_data
-
-    @staticmethod
-    def _get_box_data(b_plotter, box_name):
-        """
-        box_name can be either a name "cat" or a tuple ("cat", "hue")
-
-        Almost a duplicate of seaborn's code, because there is not
-        direct access to the box_data in the BoxPlotter class.
-        """
-        if b_plotter.plot_hues is None:
-            data = Annotator._get_box_data_without_hue(b_plotter, box_name)
-        else:
-            data = Annotator._get_box_data_with_hue(b_plotter, box_name)
-
-        box_data = remove_null(data)
-
-        return box_data
-
-    @staticmethod
-    def _get_transform_func(ax, kind):
-        """
-        Given an axis object, returns one of three possible transformation
-        functions to move between coordinate systems, depending on the value of
-        kind:
-        'data_to_ax': converts data coordinates to axes coordinates
-        'ax_to_data': converts axes coordinates to data coordinates
-        'pix_to_ax': converts pixel coordinates to axes coordinates
-        'all': return tuple of all three
-
-        This function should be called whenever axes limits are altered.
-        """
-
-        check_is_in(kind, ['data_to_ax', 'ax_to_data', 'pix_to_ax', 'all'],
-                    'kind')
-
-        if kind == 'pix_to_ax':
-            return ax.transAxes.inverted()
-
-        data_to_ax = ax.transData + ax.get_xaxis_transform().inverted()
-
-        if kind == 'data_to_ax':
-            return data_to_ax
-
-        elif kind == 'ax_to_data':
-            return data_to_ax.inverted()
-
-        else:
-            return data_to_ax, data_to_ax.inverted(), ax.transAxes.inverted()
+            self._y_stack_arr[1, :] = 1
 
     def _check_test_pvalues_perform(self):
         if self.test is None:
@@ -749,41 +526,11 @@ class Annotator:
             raise ValueError(
                 "`text_annot_custom` should be of same length as `box_pairs`.")
 
-    @staticmethod
-    def _apply_type1_comparisons_correction(comparisons_correction,
-                                            test_result_list):
-        original_pvalues = [result.pvalue for result in test_result_list]
-
-        significant_pvalues = comparisons_correction(original_pvalues)
-        correction_name = comparisons_correction.name
-        for is_significant, result in zip(significant_pvalues,
-                                          test_result_list):
-            result.correction_method = correction_name
-            result.corrected_significance = is_significant
-
-    @staticmethod
-    def _apply_type0_comparisons_correction(comparisons_correction,
-                                            test_result_list):
-        alpha = comparisons_correction.alpha
-        for result in test_result_list:
-            result.correction_method = comparisons_correction.name
-            result.corrected_significance = (
-                    result.pvalue < alpha
-                    or np.isclose(result.pvalue, alpha))
-
-    @staticmethod
-    def _apply_comparisons_correction(comparisons_correction,
-                                      test_result_list):
-        if comparisons_correction is None:
+    def _apply_comparisons_correction(self, test_result_list):
+        if self.comparisons_correction is None:
             return
-
-        if comparisons_correction.type == 1:
-            Annotator._apply_type1_comparisons_correction(
-                comparisons_correction, test_result_list)
-
         else:
-            Annotator._apply_type0_comparisons_correction(
-                comparisons_correction, test_result_list)
+            self.comparisons_correction.apply(test_result_list)
 
     def _get_stat_result_from_test(self, box_struct1, box_struct2,
                                    num_comparisons,
@@ -819,104 +566,6 @@ class Annotator:
         return result
 
     @staticmethod
-    def _get_box_struct_pairs(box_pairs, box_structs):
-        box_structs_dic = {box_struct['box']: box_struct
-                           for box_struct in box_structs}
-
-        box_struct_pairs = []
-
-        for i_box_pair, (box1, box2) in enumerate(box_pairs):
-            box_struct1 = dict(box_structs_dic[box1], i_box_pair=i_box_pair)
-            box_struct2 = dict(box_structs_dic[box2], i_box_pair=i_box_pair)
-
-            if box_struct1['x'] <= box_struct2['x']:
-                box_struct_pairs.append((box_struct1, box_struct2))
-            else:
-                box_struct_pairs.append((box_struct2, box_struct1))
-
-        return box_struct_pairs
-
-    @staticmethod
-    def absolute_box_struct_pair_in_tuple_x_diff(box_struct_pair):
-        return abs(box_struct_pair[1][1]['x'] - box_struct_pair[1][0]['x'])
-
-    def _sort_box_struct_pairs(self):
-        # Draw first the annotations with the shortest between-boxes distance,
-        # in order to reduce overlapping between annotations.
-        new_box_struct_pairs = sorted(
-            enumerate(self.box_struct_pairs),
-            key=self.absolute_box_struct_pair_in_tuple_x_diff)
-
-        self._reordering, self.box_struct_pairs = zip(*new_box_struct_pairs)
-
-    # noinspection PyProtectedMember
-    @staticmethod
-    def _get_plotter(plot, x, y, hue, data, order, hue_order,
-                     **plot_params):
-        if not plot_params.pop("dodge", True):
-            raise ValueError("`dodge` must be True in statannotations")
-
-        if plot == 'boxplot':
-            plotter = sns.categorical._BoxPlotter(
-
-                x, y, hue, data, order, hue_order,
-                orient=plot_params.get("orient"),
-                width=plot_params.get("width", 0.8),
-                dodge=True,
-                fliersize=plot_params.get("fliersize", 5),
-                linewidth=plot_params.get("linewidth"),
-                saturation=.75, color=None, palette=None)
-
-        elif plot == 'swarmplot':
-            plotter = sns.categorical._SwarmPlotter(
-                x, y, hue, data, order, hue_order,
-                orient=plot_params.get("orient"),
-                dodge=True, color=None, palette=None)
-
-        elif plot == 'stripplot':
-            plotter = sns.categorical._StripPlotter(
-                x, y, hue, data, order, hue_order,
-                jitter=plot_params.get("jitter", True),
-                orient=plot_params.get("orient"),
-                dodge=True, color=None, palette=None)
-
-        elif plot == 'barplot':
-            plotter = sns.categorical._BarPlotter(
-                x, y, hue, data, order, hue_order,
-                estimator=plot_params.get("estimator", np.mean),
-                ci=plot_params.get("ci", 95),
-                n_boot=plot_params.get("nboot", 1000),
-                units=plot_params.get("units"),
-                orient=plot_params.get("orient"),
-                seed=plot_params.get("seed"),
-                color=None, palette=None, saturation=.75,
-                errcolor=".26", errwidth=plot_params.get("errwidth"),
-                capsize=None,
-                dodge=True)
-
-        elif plot == "violinplot":
-            plotter = sns.categorical._ViolinPlotter(
-                x, y, hue, data, order, hue_order,
-                bw=plot_params.get("bw", "scott"),
-                cut=plot_params.get("cut", 2),
-                scale=plot_params.get("scale", "area"),
-                scale_hue=plot_params.get("scale_hue", True),
-                gridsize=plot_params.get("gridsize", 100),
-                width=plot_params.get("width", 0.8),
-                inner=plot_params.get("inner", None),
-                split=plot_params.get("split", False),
-                dodge=True, orient=plot_params.get("orient"),
-                linewidth=plot_params.get("linewidth"), color=None,
-                palette=None, saturation=.75)
-
-        else:
-            raise NotImplementedError(
-                f"Only {render_collection(IMPLEMENTED_PLOTTERS)} are "
-                f"supported.")
-
-        return plotter
-
-    @staticmethod
     def _get_offsets_inside(line_offset, line_offset_to_box):
         if line_offset is None:
             line_offset = 0.05
@@ -939,42 +588,6 @@ class Annotator:
 
         y_offset = line_offset
         return y_offset, line_offset_to_box
-
-    @staticmethod
-    def _get_box_structs(box_plotter, box_names, labels, ymaxes):
-        box_structs = [
-            {
-                'box': box_names[i],
-                'label': labels[i],
-                'x': Annotator._get_box_x_position(box_plotter, box_names[i]),
-                'box_data': Annotator._get_box_data(box_plotter, box_names[i]),
-                'ymax': ymaxes[box_names[i]]
-            } for i in range(len(box_names))]
-
-        # Sort the box data structures by position along the x axis
-        box_structs = sorted(box_structs, key=lambda a: a['x'])
-
-        # Add the index position in the list of boxes along the x axis
-        box_structs = [dict(box_struct, xi=i)
-                       for i, box_struct in enumerate(box_structs)]
-
-        return box_structs
-
-    @staticmethod
-    def _get_box_names_and_labels(box_plotter):
-        group_names = box_plotter.group_names
-        hue_names = box_plotter.hue_names
-
-        if box_plotter.plot_hues is None:
-            box_names = group_names
-            labels = box_names
-        else:
-            box_names = [(group_name, hue_name) for group_name in group_names
-                         for hue_name in hue_names]
-            labels = ['{}_{}'.format(group_name, hue_name)
-                      for (group_name, hue_name) in box_names]
-
-        return box_names, labels
 
     def validate_test_short_name(self):
         if self.test_short_name is not None:
@@ -1010,7 +623,7 @@ class Annotator:
         if not self.use_fixed_offset:
             try:
                 bbox = ann.get_window_extent()
-                pix_to_ax = Annotator._get_transform_func(self.ax, 'pix_to_ax')
+                pix_to_ax = self._plotter.get_transform_func('pix_to_ax')
                 bbox_ax = bbox.transformed(pix_to_ax)
                 y_top_annot = bbox_ax.ymax
 
@@ -1046,12 +659,12 @@ class Annotator:
 
     def _get_y_for_pair(self, i_ymax_in_range_x1_x2):
 
-        ymax_in_range_x1_x2 = self.y_stack_arr[1, i_ymax_in_range_x1_x2]
+        ymax_in_range_x1_x2 = self._y_stack_arr[1, i_ymax_in_range_x1_x2]
 
         # Choose the best offset depending on whether there is an annotation
         # below at the x position in the range [x1, x2] where the stack is the
         # highest
-        if self.y_stack_arr[2, i_ymax_in_range_x1_x2] == 0:
+        if self._y_stack_arr[2, i_ymax_in_range_x1_x2] == 0:
             # there is only a box below
             offset = self.line_offset_to_box
         else:
