+++ statannotations/_version.py
@@ -1,1 +1,1 @@
-__version__ = "0.4.0a"
+__version__ = "0.4.0-alpha2"
